+++ back-end/Models/followSystem/get-followers.js
@@ -1,8 +1,8 @@
 const connection = require("../../db");
 const getFollowers = (req, res) => {
-  const data = req.params.username;
+  const { username } = req.params;
   const query = `SELECT following_username FROM Follower_system WHERE follower_username =?`;
-  connection.query(query, data, (err, results) => {
+  connection.query(query, username, (err, results) => {
     if (err) throw err.sqlMessage;
 
     res.json(results);
