+++ back-end/Models/followSystem/get-following.js
@@ -1,8 +1,8 @@
 const connection = require("../../db");
 const getUserFollowing = (req, res) => {
-  const data = req.params.username;
+  const { username } = req.params;
   const query = `SELECT follower_username FROM Follower_system WHERE following_username =?`;
-  connection.query(query, data, (err, results) => {
+  connection.query(query, username, (err, results) => {
     if (err) throw err.sqlMessage;
     res.json(results);
   });
