+++ back-end/Models/followSystem/unfollows.js
@@ -1,11 +1,8 @@
-const jwt_decode = require("jwt-decode");
 const connection = require("../../db");
 const unfollows = (req, res) => {
-  const token = req.headers.authorization.split(" ").pop();
-  const decoded = jwt_decode(token);
-  const followerData = [decoded.username, req.params.username];
-  const followerQuery = `DELETE FROM Follower_system WHERE (follower_username,following_username) = (?,?)`;
-  connection.query(followerQuery, followerData, (err, result) => {
+  const data = [req.token.username, req.params.username];
+  const query = `DELETE FROM Follower_system WHERE (follower_username,following_username) = (?,?)`;
+  connection.query(query, data, (err, result) => {
     if (err) throw err.sqlMessage;
   });
 };
