+++ back-end/package.json
@@ -16,7 +16,10 @@
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^3.1.2",
-    "mysql2": "^2.2.5"
+    "mysql2": "^2.2.5",
+    "socket.io": "^3.0.3",
+    "uuid": "^8.3.1",
+    "uuidv4": "^6.2.5"
   },
   "devDependencies": {},
   "description": ""
